+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Unreleased
 
+- [#671](https://github.com/squint-cljs/squint/issues/671): Implement `trampoline`
 - Fix [#673](https://github.com/squint-cljs/squint/issues/673): remove experimental atom as promise option since it causes unexpected behavior
 
 ## v0.8.148